Implémentation du bouton de suppression du textfield Calcul des conversions $>€, de la TVA et des pourboires Sauvegarde des valeurs d’entrées et de sorties de l’UI pour les conversions $>€, la TVA et les pourboires

+++ MonBaluchonParisNYC/Model/Currency/CurrencyRateHttpData.swift
@@ -13,6 +13,6 @@ struct CurrencyRateHttpData: Decodable {
     let rates: Rates
     
     struct Rates: Decodable {
-        let USD: Float
+        let USD: Double
     }
 }
